requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2==2.9.11
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
